refactor: rename common.extensions to common.extload
More prep-work for the extension API.

This module is the extension loading code.  This renames it to
`extload` so as not to confuse it with the future extension API,
which will end up in `alpenhorn.extensions`.

There are no code changes here.
